\name{unival-internal}
\alias{unival-internal}
\alias{univalboot}
\alias{univalboot2}
\alias{size}
\alias{transpose}
\title{
unival-internal
}
\description{
Functions for internal use. Most precisely, those are used in the computation of the bootstrap analysis.
}
\usage{
univalboot
univalboot2
size
transpose
}
\author{
Pere Joan Ferrando

David Navarro-Gonzalez

Urbano Lorenzo-Seva
}
\keyword{ internal }
